+++ src/models/Election/ElectionBaseDTO.ts
